[Feature] Merge upstream 105.0.5144.1

+++ base/android/java/src/org/chromium/base/FileUtils.java
@@ -9,8 +9,12 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.ParcelFileDescriptor;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+import org.chromium.base.annotations.JNINamespace;
+import org.chromium.base.annotations.NativeMethods;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -25,23 +29,38 @@ import java.util.Locale;
 /**
  * Helper methods for dealing with Files.
  */
+@JNINamespace("base::android")
 public class FileUtils {
     private static final String TAG = "FileUtils";
+
+    public static Function<String, Boolean> DELETE_ALL = filepath -> true;
 
     /**
      * Delete the given File and (if it's a directory) everything within it.
      * @param currentFile The file or directory to delete. Does not need to exist.
-     * @return Whether currentFile does not exist afterwards.
+     * @param canDelete the {@link Function} function used to check if the file can be deleted.
+     * @return True if the files are deleted, or files reserved by |canDelete|, false if failed to
+     *         delete files.
+     * @note Caveat: Return values from recursive deletes are ignored.
+     * @note Caveat: |canDelete| is not robust; see https://crbug.com/1066733.
      */
-    public static boolean recursivelyDeleteFile(File currentFile) {
+    public static boolean recursivelyDeleteFile(
+            File currentFile, Function<String, Boolean> canDelete) {
         if (!currentFile.exists()) {
+            // This file could be a broken symlink, so try to delete. If we don't delete a broken
+            // symlink, the directory containing it cannot be deleted.
+            currentFile.delete();
             return true;
         }
+        if (canDelete != null && !canDelete.apply(currentFile.getPath())) {
+            return true;
+        }
+
         if (currentFile.isDirectory()) {
             File[] files = currentFile.listFiles();
             if (files != null) {
                 for (File file : files) {
-                    recursivelyDeleteFile(file);
+                    recursivelyDeleteFile(file, canDelete);
                 }
             }
         }
@@ -57,31 +76,40 @@ public class FileUtils {
      * Delete the given files or directories by calling {@link #recursivelyDeleteFile(File)}. This
      * supports deletion of content URIs.
      * @param filePaths The file paths or content URIs to delete.
+     * @param canDelete the {@link Function} function used to check if the file can be deleted.
      */
-    public static void batchDeleteFiles(List<String> filePaths) {
+    public static void batchDeleteFiles(
+            List<String> filePaths, Function<String, Boolean> canDelete) {
         for (String filePath : filePaths) {
+            if (canDelete != null && !canDelete.apply(filePath)) continue;
             if (ContentUriUtils.isContentUri(filePath)) {
                 ContentUriUtils.delete(filePath);
             } else {
                 File file = new File(filePath);
-                if (file.exists()) recursivelyDeleteFile(file);
+                if (file.exists()) recursivelyDeleteFile(file, canDelete);
             }
         }
     }
 
     /**
-     * Extracts an asset from the app's APK to a file.
-     * @param context
-     * @param assetName Name of the asset to extract.
-     * @param outFile File to extract the asset to.
-     * @return true on success.
+     * Get file size. If it is a directory, recursively get the size of all files within it.
+     * @param file The file or directory.
+     * @return The size in bytes.
      */
-    public static boolean extractAsset(Context context, String assetName, File outFile) {
-        try (InputStream inputStream = context.getAssets().open(assetName)) {
-            copyStreamToFile(inputStream, outFile);
-            return true;
-        } catch (IOException e) {
-            return false;
+    public static long getFileSizeBytes(File file) {
+        if (file == null) return 0L;
+        if (file.isDirectory()) {
+            long size = 0L;
+            final File[] files = file.listFiles();
+            if (files == null) {
+                return size;
+            }
+            for (File f : files) {
+                size += getFileSizeBytes(f);
+            }
+            return size;
+        } else {
+            return file.length();
         }
     }
 
@@ -149,13 +177,14 @@ public class FileUtils {
 
     /**
      * Returns the file extension, or an empty string if none.
-     * @param file Name of the file, with or without the full path.
+     * @param file Name of the file, with or without the full path (Unix style).
      * @return empty string if no extension, extension otherwise.
      */
     public static String getExtension(String file) {
-        int index = file.lastIndexOf('.');
-        if (index == -1) return "";
-        return file.substring(index + 1).toLowerCase(Locale.US);
+        int lastSep = file.lastIndexOf('/');
+        int lastDot = file.lastIndexOf('.');
+        if (lastSep >= lastDot) return ""; // Subsumes |lastDot == -1|.
+        return file.substring(lastDot + 1).toLowerCase(Locale.US);
     }
 
     /** Queries and decodes bitmap from content provider. */
@@ -182,5 +211,23 @@ public class FileUtils {
             Log.w(TAG, "IO exception when reading uri " + uri);
         }
         return null;
+    }
+
+    /**
+     * Gets the canonicalised absolute pathname for |filePath|. Returns empty string if the path is
+     * invalid. This function can result in I/O so it can be slow.
+     * @param filePath Path of the file, has to be a file path instead of a content URI.
+     * @return canonicalised absolute pathname for |filePath|.
+     */
+    public static String getAbsoluteFilePath(String filePath) {
+        return FileUtilsJni.get().getAbsoluteFilePath(filePath);
+    }
+
+    @NativeMethods
+    public interface Natives {
+        /**
+         * Returns the canonicalised absolute pathname for |filePath|.
+         */
+        String getAbsoluteFilePath(String filePath);
     }
 }
