+++ chrome/android/javatests/src/org/chromium/chrome/browser/download/DownloadTestRule.java
@@ -7,6 +7,8 @@ package org.chromium.chrome.browser.download;
 import android.app.DownloadManager;
 import android.content.Context;
 import android.database.Cursor;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Environment;
 import android.text.TextUtils;
 
@@ -17,10 +19,10 @@ import org.junit.runners.model.Statement;
 import org.chromium.base.ApiCompatibilityUtils;
 import org.chromium.base.Log;
 import org.chromium.base.test.util.CallbackHelper;
-import org.chromium.chrome.browser.ChromeActivity;
 import org.chromium.chrome.browser.download.items.OfflineContentAggregatorFactory;
-import org.chromium.chrome.browser.preferences.PrefServiceBridge;
-import org.chromium.chrome.test.ChromeActivityTestRule;
+import org.chromium.chrome.browser.profiles.ProfileKey;
+import org.chromium.chrome.test.ChromeTabbedActivityTestRule;
+import org.chromium.components.download.DownloadCollectionBridge;
 import org.chromium.components.offline_items_collection.ContentId;
 import org.chromium.components.offline_items_collection.OfflineContentProvider;
 import org.chromium.components.offline_items_collection.OfflineItem;
@@ -31,7 +33,6 @@ import org.chromium.content_public.browser.test.util.TestThreadUtils;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -45,16 +46,16 @@ import java.util.concurrent.TimeoutException;
  * and all of our test cases.
  *
  */
-public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
+public class DownloadTestRule extends ChromeTabbedActivityTestRule {
     private static final String TAG = "DownloadTestBase";
     private static final File DOWNLOAD_DIRECTORY =
             Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
     public static final long UPDATE_DELAY_MILLIS = 1000;
 
     private final CustomMainActivityStart mActivityStart;
+    private List<DownloadItem> mAllDownloads;
 
     public DownloadTestRule(CustomMainActivityStart action) {
-        super(ChromeActivity.class);
         mActivityStart = action;
     }
 
@@ -177,7 +178,7 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
         return mHttpDownloadFinished.getCallCount();
     }
 
-    public boolean waitForChromeDownloadToFinish(int currentCallCount) throws InterruptedException {
+    public boolean waitForChromeDownloadToFinish(int currentCallCount) {
         boolean eventReceived = true;
         try {
             mHttpDownloadFinished.waitForCallback(currentCallCount, 1, 5, TimeUnit.SECONDS);
@@ -187,17 +188,25 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
         return eventReceived;
     }
 
+    public List<DownloadItem> getAllDownloads() {
+        TestThreadUtils.runOnUiThreadBlocking(() -> {
+            DownloadManagerService.getDownloadManagerService().getAllDownloads(null);
+        });
+        return mAllDownloads;
+    }
+
     private class TestDownloadManagerServiceObserver
             implements DownloadManagerService.DownloadObserver {
         @Override
-        public void onAllDownloadsRetrieved(final List<DownloadItem> list, boolean isOffTheRecord) {
+        public void onAllDownloadsRetrieved(final List<DownloadItem> list, ProfileKey profileKey) {
+            mAllDownloads = list;
         }
 
         @Override
         public void onDownloadItemCreated(DownloadItem item) {}
 
         @Override
-        public void onDownloadItemRemoved(String guid, boolean isOffTheRecord) {}
+        public void onDownloadItemRemoved(String guid) {}
 
         @Override
         public void onAddOrReplaceDownloadSharedPreferenceEntry(ContentId id) {}
@@ -206,7 +215,7 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
         public void onDownloadItemUpdated(DownloadItem item) {}
 
         @Override
-        public void broadcastDownloadSuccessfulForTesting(DownloadInfo downloadInfo) {
+        public void broadcastDownloadSuccessful(DownloadInfo downloadInfo) {
             mLastDownloadFilePath = downloadInfo.getFilePath();
             mHttpDownloadFinished.notifyCalled();
         }
@@ -214,7 +223,7 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
 
     private class TestDownloadBackendObserver implements OfflineContentProvider.Observer {
         @Override
-        public void onItemsAdded(ArrayList<OfflineItem> items) {}
+        public void onItemsAdded(List<OfflineItem> items) {}
 
         @Override
         public void onItemRemoved(ContentId id) {}
@@ -244,8 +253,7 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
         mActivityStart.customMainActivityStart();
 
         TestThreadUtils.runOnUiThreadBlocking(() -> {
-            PrefServiceBridge.getInstance().setPromptForDownloadAndroid(
-                    DownloadPromptStatus.DONT_SHOW);
+            DownloadDialogBridge.setPromptForDownloadAndroid(DownloadPromptStatus.DONT_SHOW);
         });
 
         cleanUpAllDownloads();
@@ -258,7 +266,7 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
         });
     }
 
-    private void tearDown() throws Exception {
+    private void tearDown() {
         cleanUpAllDownloads();
         TestThreadUtils.runOnUiThreadBlocking(() -> {
             DownloadManagerService.getDownloadManagerService().removeDownloadObserver(
@@ -267,13 +275,27 @@ public class DownloadTestRule extends ChromeActivityTestRule<ChromeActivity> {
     }
 
     public void deleteFilesInDownloadDirectory(String... filenames) {
-        for (String filename : filenames) {
-            final File fileToDelete = new File(DOWNLOAD_DIRECTORY, filename);
+        for (String filename : filenames) deleteFile(filename);
+    }
+
+    private void deleteFile(String fileName) {
+        // Delete file path on pre Q.
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+            final File fileToDelete = new File(DOWNLOAD_DIRECTORY, fileName);
             if (fileToDelete.exists()) {
                 Assert.assertTrue(
                         "Could not delete file that would block this test", fileToDelete.delete());
             }
+            return;
         }
+
+        // Delete content URI starting from Q.
+        Uri uri = DownloadCollectionBridge.getDownloadUriForFileName(fileName);
+        if (uri == null) {
+            Log.e(TAG, "Can't find URI of file for deletion: %s on Android P+.", fileName);
+            return;
+        }
+        DownloadCollectionBridge.deleteIntermediateUri(uri.toString());
     }
 
     /**
